* | Ton Balance | visiable state ton balance view

+++ Sources/Common/Network/Models/TonClient.swift
@@ -3,10 +3,10 @@ import Foundation
 public struct TonClient: Codable {
     let clientId: String
     let address: String
-    var balance: Float
-    var usd: Float
+    var balance: Double
+    var usd: Double
     
-    init(clientId: String, address: String, balance: Float = 0.0, usd: Float = 0.0) {
+    init(clientId: String, address: String, balance: Double = 0.0, usd: Double = 0.0) {
         self.clientId = clientId
         self.address = address
         self.balance = balance
